// ui/src/main/angular/ui/modules/lipidcompass-backend-client/model/entityModelLipid.ts
/**
 * LipidCompass Backend API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { FattyAcyl } from './fattyAcyl';
import { LipidMapsEntry } from './lipidMapsEntry';
import { CrossReference } from './crossReference';
import { Link } from './link';
import { SwissLipidsEntry } from './swissLipidsEntry';


export interface EntityModelLipid { 
    id?: string;
    arangoId?: string;
    transactionUuid: string;
    visibility?: EntityModelLipid.VisibilityEnum;
    nativeUrl?: string;
    commonName?: string;
    normalizedShorthandName?: string;
    lipidLevel?: EntityModelLipid.LipidLevelEnum;
    synonyms?: Array<string>;
    systematicName?: string;
    chemicalFormula?: string;
    exactMass?: number;
    inchiKey?: string;
    inchi?: string;
    smiles?: string;
    mdlModel?: string;
    crossReferences?: Array<CrossReference>;
    lipidCategory?: string;
    lipidClass?: string;
    lipidSpecies?: string;
    lipidMolecularSpecies?: string;
    lipidSnPosition?: string;
    lipidStructureDefined?: string;
    lipidFullStructure?: string;
    lipidCompleteStructure?: string;
    nativeId?: string;
    lipidSpeciesInfo?: FattyAcyl;
    lipidMapsEntry?: Array<LipidMapsEntry>;
    swissLipidsEntry?: Array<SwissLipidsEntry>;
    fattyAcyls?: Array<FattyAcyl>;
    revision?: string;
    dateCreated?: string;
    dateLastModified?: string;
    createdBy?: string;
    updatedBy?: string;
    _links?: { [key: string]: Link; };
}
export namespace EntityModelLipid {
    export type VisibilityEnum = 'PRIVATE' | 'RESTRICTED' | 'PUBLIC' | '11184809';
    export const VisibilityEnum = {
        PRIVATE: 'PRIVATE' as VisibilityEnum,
        RESTRICTED: 'RESTRICTED' as VisibilityEnum,
        PUBLIC: 'PUBLIC' as VisibilityEnum,
        unknown_default_open_api: '11184809' as VisibilityEnum
    };
    export type LipidLevelEnum = 'NO_LEVEL' | 'UNDEFINED_LEVEL' | 'CATEGORY' | 'CLASS' | 'SPECIES' | 'MOLECULAR_SPECIES' | 'SN_POSITION' | 'STRUCTURE_DEFINED' | 'FULL_STRUCTURE' | 'COMPLETE_STRUCTURE' | '11184809';
    export const LipidLevelEnum = {
        NO_LEVEL: 'NO_LEVEL' as LipidLevelEnum,
        UNDEFINED_LEVEL: 'UNDEFINED_LEVEL' as LipidLevelEnum,
        CATEGORY: 'CATEGORY' as LipidLevelEnum,
        CLASS: 'CLASS' as LipidLevelEnum,
        SPECIES: 'SPECIES' as LipidLevelEnum,
        MOLECULAR_SPECIES: 'MOLECULAR_SPECIES' as LipidLevelEnum,
        SN_POSITION: 'SN_POSITION' as LipidLevelEnum,
        STRUCTURE_DEFINED: 'STRUCTURE_DEFINED' as LipidLevelEnum,
        FULL_STRUCTURE: 'FULL_STRUCTURE' as LipidLevelEnum,
        COMPLETE_STRUCTURE: 'COMPLETE_STRUCTURE' as LipidLevelEnum,
        unknown_default_open_api: '11184809' as LipidLevelEnum
    };
}