// ui/src/main/angular/ui/modules/lipidcompass-backend-client/model/pageMetadata.ts
/**
 * LipidCompass Backend API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PageMetadata { 
    size?: number;
    totalElements?: number;
    totalPages?: number;
    number?: number;
}